requirements.txt
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.1
beautifulsoup4==4.14.2
lxml==4.9.3
